reorganized project folder and files - noew divided in modules
