+++ client/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { LandingPageComponent } from './landing-page/landing-page.component';
 import { AuthComponent } from './auth.module/auth/auth.component';
-import { BudgetsPageComponent } from './budgets-page/budgets-page.component';
+import { BudgetsPageComponent } from './budget-module/budgets-page/budgets-page.component';
 import { LoginComponent } from './auth.module/login/login.component';
 import { SignupComponent } from './auth.module/signup/signup.component';
 import { AppComponent } from './app.component';
