+++ client/app/auth.module/auth.module.ts
@@ -1,0 +1,23 @@
+import { RouterModule } from '@angular/router';
+import { SignupComponent } from './signup/signup.component';
+import { LoginComponent } from './login/login.component';
+import { ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+
+@NgModule({
+    declarations: [
+        LoginComponent,
+        SignupComponent,
+    ],
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        RouterModule
+    ],
+    providers: [
+
+    ]
+})
+export class AuthModule{}
