+++ client/app/budget-module/budget-overview/budget-overview.component.ts
@@ -1,4 +1,4 @@
-import { Budget } from './../models/budget.model';
+import { Budget } from './../../models/budget.model';
 import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
