+++ client/app/budget-module/navbar/navbar.component.scss
